Add an "installVersion" concept to the package.json.
This is the "version" of the dev header files that are installed. When
`node-gyp install --ensure` is used, if the installed versions "installVersion"
doesn't match what node-gyp is expecting here, then it will re-install.

Increment this number any time the installation procedure in the "install"
phase changes, and thus the header files needing to be reinstalled.