Update README.md
There is a known issue with the Windows SDK and Visual Studio. It seems that when both are installed users may be missing the 64 bit compilers which will make using node-gyp for 64 bit really frustrating. It took me quite a bit of googling to find the information about this issue, as it didn't come up when searching for the various error messages and such. I've added a note to the readme to make it easier for people to find the compiler update and get node-gyp working properly on 64 bit systems with this common issue. :D

see the kb article: http://support.microsoft.com/?kbid=2519277 for some details. I had Visual Studio 2010 and 2012 fully updated before getting the SDK, the article describes a different scenario which leads to the same problem.