Handle when both VS2010 and VS2012 are installed
When both versions are present, attempt to use the command prompt to
determine which one should be set.

The command prompt for VS2012 set the environment variable to:

VisualStudioVersion=11.0

This environment variable is absent in the VS2010 command prompts.

Also, presumably the checkVCExpress function not setting hasVCExpress
was an oversight.