Allow glob@4
There are no actual breaking changes in glob 4.

The only breaking change is that it requires a new version of npm,
because it uses the ^ syntax to specify its dependency.

This is not a problem for node-gyp, because new versions of node-gyp
will generally come along with new versions of node and/or npm.