gyp: fix runtime "NameError"s
In "gyp/gyptest.py", `shlex` is used, but it is not imported. So it
will throw `NameError` at runtime.

In "gyp/pylib/gyp/mac_tool.py", the `except` block expects the type of
exception to be the first item or nothing. As it is, the block will
throw a `NameError` at runtime.