Prefix build targets with /t: on Windows
Currently, on non-Windows platforms, it is possible to have
a multi-target native module and specify building just some
of the targets using `node-gyp build my_target`.

On Windows, however, specifying the targets to build requires the `/t:`
or `/target:` flag. Without this change you will get an `MSB1008` error
because MSBuild thinks you are trying to specify multiple solutions.

PR-URL: https://github.com/nodejs/node-gyp/pull/1164
Reviewed-By: Ben Noordhuis <info@bnoordhuis.nl>