build,win: use config.gypi MSBuild path over env
When vs2017 and vs2015 are both installed and vs2015's MSBuild.exe
is in the Path environment variable, node-gyp fails to build if it
configured the project to build with vs2017, because it tries to use
vs2015's MSBuild.exe, which is unable to find the v141 (vs2017)
Platform Toolset build tools.

Change build behaviour to prefer using the MSBuild.exe path if it is
defined in config.gypi instead of using the MSBuild.exe found in Path
by "which".

Ref: https://github.com/nodejs/node/issues/13641