test: use Travis CI to run tests on every pull request
This is a second attempt at #1336 which got into a bad git-state...

Use flake8 to find Python syntax errors and undefined names.  There are Python 3 syntax errors and many undefined names which may raise NameError at runtime. This PR runs flake8 runs in two passes: The first looks at critical issues in stop-the-build mode and the second looks at style violations in everything-is-a-warning mode.

PR-URL: https://github.com/nodejs/node-gyp/pull/1752
Reviewed-By: Rod Vagg <rod@vagg.org>