doc, bin: stop suggesting opening  node-gyp issues
A lot of new issues in node-gyp are related to outdated node-gyp or
broken modules. This removes the suggestion to open a new issue in the
node-gyp, instead suggesting the user should open the issue in the
module issue tracker. It also makes the issue template more explicit
about providing the logs.

PR-URL: https://github.com/nodejs/node-gyp/pull/2096
Reviewed-By: Christian Clauss <cclauss@me.com>
Reviewed-By: Rod Vagg <rod@vagg.org>