doc, bin: stop suggesting opening  node-gyp issues
A lot of new issues in node-gyp are related to outdated node-gyp or
broken modules. This removes the suggestion to open a new issue in the
node-gyp, instead suggesting the user should open the issue in the
module issue tracker. It also makes the issue template more explicit
about providing the logs.