ci: add .github/workflows/release-please.yml
Automatically creates/updates release PRs when there are new commits
on a given branch of the repository (currently configured as `master`).

Semver hints in commit messages can let release-please know
whether to draft a minor or major release, but otherwise the
release proposal PR will be for a semver patch-level release by default.

Co-authored-by: gengjiawen <technicalcute@gmail.com>