docs: Add notes/disclaimers for upgrading the node-gyp npm uses
Add notes on npm version compatibility for methods to update node-gyp.

Also mention that this is tested on npm 8, now that npm 8 exists.
(Npm 8 is the same as npm 7 for purposes of these instructions.)

Lastly, add a note that one of the methods is simpler/more durable,
if only functional for npm 6 and older.