fix: create Python symlink only during builds, and clean it up after
Previously in b9ddcd5bbd93b05b03674836b6ebdae2c2e74c8c this was created
during configuration, and the symlink persisted indefinitely. This
causes problems with many tools that do not expect a codebase to include
symlinks to external absolute paths.

This PR largely reverts that commit, and instead writes the path to
link to into the config, and then creates the symlink only temporarily
during the build process, always deleting it afterwards.