doc: Add note about Python symlinks (PR 2362) to CHANGELOG.md for 9.1.0 (#2783)
The PR for this change was merged without a prefixed name,
such as "lib:" or "fix:". That means release-please
didn't include it in the changelog for v9.1.0.

This change did end up affecting users, though. (See issue 2713
and PR 2721). Therefore, I believe it should be noted
in the CHANGELOG.md, so users can better understand the behavior
they are seeing.