fix: disable msbuild.exe nodeReuse
On windows, one may experience errors due to concurrent node-gyp calls (see #3095)

When retrying after such a failure, retries are observed to fail repeatedly with strange
EBUSY and EPERM errors. This turns out to be due to a caching feature where msbuild.exe
will continue to run in the background (for either 15s or 15m depending on VS version),
and keep open file handles to files and directories that the user may be trying to delete
(for example with a new call to `npm ci`).

This behavior is well documented, as is the recommended workaround implemented here.